+++ grails-app/domain/com/causecode/currency/Currency.groovy
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2011, CauseCode Technologies Pvt Ltd, India.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or
+ * without modification, are not permitted.
+ */
+
+package com.causecode.currency
+
+/**
+ * Used to store currency code and name information.
+ *
+ */
+
+class Currency {
+
+    Date dateCreated
+    Date lastUpdated
+
+    String code
+    String name
+
+    static constraints = {
+        dateCreated bindable: false
+        lastUpdated bindable: false
+        code blank: false
+        name blank: false
+    }
+
+    @Override
+    String toString() {
+        code
+    }
+
+}