+++ grails-app/domain/com/causecode/user/Role.groovy
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 2011, CauseCode Technologies Pvt Ltd, India.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or
+ * without modification, are not permitted.
+ */
+
+package com.causecode.user
+
+/**
+ * Role groovy class used to specify authority information.
+ */
+class Role {
+
+    String authority
+
+    static mapping = {
+        cache true
+    }
+
+    static constraints = {
+        authority blank: false, unique: true
+    }
+
+}